Move autocomplete endpoints to separate file in worksearch
