+++ openlibrary/plugins/upstream/addbook.py
@@ -1,7 +1,6 @@
 """Handlers for adding and editing books."""
 
 import io
-import itertools
 import web
 import json
 import csv
@@ -10,15 +9,12 @@ import datetime
 from typing import Literal, overload, NoReturn
 
 from infogami import config
-from infogami.core import code as core
 from infogami.core.db import ValidationException
 from infogami.utils import delegate
 from infogami.utils.view import safeint, add_flash_message
 from infogami.infobase.client import ClientException
 
-from openlibrary.plugins.openlibrary.processors import urlsafe
 from openlibrary.plugins.worksearch.search import get_solr
-from openlibrary.utils import find_author_olid_in_string, find_work_olid_in_string
 from openlibrary.i18n import gettext as _
 from openlibrary import accounts
 import logging
@@ -1033,140 +1029,6 @@ class daisy(delegate.page):
             raise web.notfound()
 
         return render_template("books/daisy", page)
-
-
-def to_json(d):
-    web.header('Content-Type', 'application/json')
-    return delegate.RawText(json.dumps(d))
-
-
-class languages_autocomplete(delegate.page):
-    path = "/languages/_autocomplete"
-
-    def GET(self):
-        i = web.input(q="", limit=5)
-        i.limit = safeint(i.limit, 5)
-        return to_json(
-            list(itertools.islice(utils.autocomplete_languages(i.q), i.limit))
-        )
-
-
-class works_autocomplete(delegate.page):
-    path = "/works/_autocomplete"
-
-    def GET(self):
-        i = web.input(q="", limit=5)
-        i.limit = safeint(i.limit, 5)
-
-        solr = get_solr()
-
-        # look for ID in query string here
-        q = solr.escape(i.q).strip()
-        embedded_olid = find_work_olid_in_string(q)
-        if embedded_olid:
-            solr_q = 'key:"/works/%s"' % embedded_olid
-        else:
-            solr_q = f'title:"{q}"^2 OR title:({q}*)'
-
-        params = {
-            'q_op': 'AND',
-            'sort': 'edition_count desc',
-            'rows': i.limit,
-            'fq': 'type:work',
-            # limit the fields returned for better performance
-            'fl': 'key,title,subtitle,cover_i,first_publish_year,author_name,edition_count',
-        }
-
-        data = solr.select(solr_q, **params)
-        # exclude fake works that actually have an edition key
-        docs = [d for d in data['docs'] if d['key'][-1] == 'W']
-
-        if embedded_olid and not docs:
-            # Grumble! Work not in solr yet. Create a dummy.
-            key = '/works/%s' % embedded_olid
-            work = web.ctx.site.get(key)
-            if work:
-                docs = [work.as_fake_solr_record()]
-
-        for d in docs:
-            # Required by the frontend
-            d['name'] = d['key'].split('/')[-1]
-            d['full_title'] = d['title']
-            if 'subtitle' in d:
-                d['full_title'] += ": " + d['subtitle']
-
-        return to_json(docs)
-
-
-class authors_autocomplete(delegate.page):
-    path = "/authors/_autocomplete"
-
-    def GET(self):
-        i = web.input(q="", limit=5)
-        i.limit = safeint(i.limit, 5)
-
-        solr = get_solr()
-
-        q = solr.escape(i.q).strip()
-        embedded_olid = find_author_olid_in_string(q)
-        if embedded_olid:
-            solr_q = 'key:"/authors/%s"' % embedded_olid
-        else:
-            prefix_q = q + "*"
-            solr_q = f'name:({prefix_q}) OR alternate_names:({prefix_q})'
-
-        params = {
-            'q_op': 'AND',
-            'sort': 'work_count desc',
-            'rows': i.limit,
-            'fq': 'type:author',
-        }
-
-        data = solr.select(solr_q, **params)
-        docs = data['docs']
-
-        if embedded_olid and not docs:
-            # Grumble! Must be a new author. Fetch from db, and build a "fake" solr resp
-            key = '/authors/%s' % embedded_olid
-            author = web.ctx.site.get(key)
-            if author:
-                docs = [author.as_fake_solr_record()]
-
-        for d in docs:
-            if 'top_work' in d:
-                d['works'] = [d.pop('top_work')]
-            else:
-                d['works'] = []
-            d['subjects'] = d.pop('top_subjects', [])
-
-        return to_json(docs)
-
-
-class subjects_autocomplete(delegate.page):
-    path = "/subjects_autocomplete"
-    # can't use /subjects/_autocomplete because the subjects endpoint = /subjects/[^/]+
-
-    def GET(self):
-        i = web.input(q="", type="", limit=5)
-        i.limit = safeint(i.limit, 5)
-
-        solr = get_solr()
-        prefix_q = solr.escape(i.q).strip()
-        solr_q = f'name:({prefix_q}*)'
-        fq = f'type:subject AND subject_type:{i.type}' if i.type else 'type:subject'
-
-        params = {
-            'fl': 'key,name,subject_type,work_count',
-            'q_op': 'AND',
-            'fq': fq,
-            'sort': 'work_count desc',
-            'rows': i.limit,
-        }
-
-        data = solr.select(solr_q, **params)
-        docs = [{'key': d['key'], 'name': d['name']} for d in data['docs']]
-
-        return to_json(docs)
 
 
 class work_identifiers(delegate.view):
